Moved components into a components folder and fixed dependencies accordingly.
